refactor: restore sqlite query file split
